refactor: prometheus packages
