+++ app/controlplane/pkg/metrics/prometheus/collector/collector.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package prometheus
+package collector
 
 import (
 	"context"
